requirements.txt
streamlit
PyMuPDF
transformers
sentencepiece
accelerate
faiss-cpu
sentence-transformers
torch
